requirements.txt
fastapi
uvicorn
httpx
pydantic
pydub
moviepy
numpy
Pillow
